break apart homescreen aggregator
