+++ aggregators/can-load-offline/index.js
