+++ index.js
@@ -50,8 +50,10 @@ const audits = [
 ];
 const aggregators = [
   require('./aggregators/will-get-add-to-homescreen-prompt'),
+  require('./aggregators/can-load-offline'),
   require('./aggregators/is-secure'),
-  require('./aggregators/is-performant')
+  require('./aggregators/is-performant'),
+  require('./aggregators/is-sized-for-mobile-screen')
 ];
 module.exports = function(opts) {
   const url = opts.url;
